backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cloudinary==1.33.0
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
